% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpm.r
\name{dpm}
\alias{dpm}
\title{Calculates the Projection Median (see Durocher 2009)}
\usage{
dpm(data, N = 10000, method = 1, exact2D = F)
}
\arguments{
\item{data}{a nxd matrix of your data. For two dimensions, d must be 2}

\item{N}{Number of vectors to generate in the Monte Carlo algorithm}

\item{method}{Which Monte Carlo algorithm to use? As per Durocher et. al 2017}

\item{exact2D}{calculate the pm exactly *data must be 2 dimensions}
}
\description{
This function allows you to calculate the projection median approximately or exactly in 2D.
}
\examples{
data<-cbind(stats::rnorm(151),stats::rnorm(151))
dpm(data)
dpm(data,exact2D=TRUE)
}
\keyword{location}
\keyword{multivariate,}
\keyword{robust}
