\name{DurocherProjectionMedian-package}
\alias{DurocherProjectionMedian-package}
\alias{DurocherProjectionMedian}
\docType{package}
\title{
\packageTitle{DurocherProjectionMedian}
}
\description{
\packageDescription{DurocherProjectionMedian}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{DurocherProjectionMedian}
\packageIndices{DurocherProjectionMedian}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{DurocherProjectionMedian}
Maintainer: \packageMaintainer{DurocherProjectionMedian}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{package}
\examples{###}
