% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionMedian2D.R
\name{projectionMedian2D}
\alias{projectionMedian2D}
\title{Calculates the Projection Median exactly in 2D (see Durocher 2009)}
\usage{
projectionMedian2D(data)
}
\arguments{
\item{data}{a nx2 matrix of your data.}
}
\description{
This function allows you to calculate the projection median exactly in 2D. Internal function for dpm()
}
\keyword{location}
\keyword{multivariate,}
\keyword{robust,}
