% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionMedianMC.R
\name{projectionMedianMC}
\alias{projectionMedianMC}
\title{Calculates the Projection Median using the Monte Carlo approximations in Durocher et. al 2017}
\usage{
projectionMedianMC(data, N = 10000, method = 1)
}
\arguments{
\item{data}{a nxd matrix of your data.}

\item{N}{Number of vectors to generate in the Monte Carlo algorithm}

\item{method}{Which Monte Carlo algorithm to use? As per Durocher et. al 2017}
}
\description{
This function allows you to calculate the projection median approximately or exactly in 2D.
}
\examples{

}
\keyword{Carlo}
\keyword{Monte}
\keyword{location,}
\keyword{multivariate,}
\keyword{robust,}
